funcionamiento-logica

+++ MiFragancia/MiFragancia/Models/Producto.cs
@@ -14,7 +14,7 @@ namespace MiFragancia.Models
         public string Nombre { get; set; }
         public string Descripcion { get; set; }
         public int Cantidad { get; set; }
-        public float Precio { get; set; }
+        public double Precio { get; set; }
         public bool Activo { get; set; }
         public int TipoProductoId { get; set; }
         [ForeignKey("TipoProductoId ")]
